requirements-core.txt
flask==3.0.3
flask-cors==4.0.1
kafka-python==2.0.2
pytest==8.3.2
